+++ src/main/java/fun/mike/flapjack/pipeline/lab/TransformPipelineError.java
@@ -3,13 +3,13 @@ package fun.mike.flapjack.pipeline.lab;
 import fun.mike.record.alpha.Record;
 
 public class TransformPipelineError implements PipelineError {
-    private final Long number;
+    private final int number;
     private final String line;
     private final Record record;
     private final OperationInfo operationInfo;
     private final Exception exception;
 
-    public TransformPipelineError(Long number, String line, Record record, OperationInfo operationInfo, Exception exception) {
+    public TransformPipelineError(int number, String line, Record record, OperationInfo operationInfo, Exception exception) {
         this.number = number;
         this.line = line;
         this.record = record;
@@ -17,11 +17,11 @@ public class TransformPipelineError implements PipelineError {
         this.exception = exception;
     }
 
-    public static TransformPipelineError of(Long number, String line, Record record, OperationInfo operationInfo, Exception exception) {
+    public static TransformPipelineError of(int number, String line, Record record, OperationInfo operationInfo, Exception exception) {
         return new TransformPipelineError(number, line, record, operationInfo, exception);
     }
 
-    public static TransformPipelineError fromResult(Long number, String line, TransformResult result) {
+    public static TransformPipelineError fromResult(int number, String line, TransformResult result) {
         return new TransformPipelineError(number, line, result.getRecord(), result.getOperationInfo(), result.getException());
     }
 
@@ -30,7 +30,7 @@ public class TransformPipelineError implements PipelineError {
     }
 
     @Override
-    public Long getNumber() {
+    public int getNumber() {
         return number;
     }
